% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_palette_class.R
\name{format.pq_palette}
\alias{format.pq_palette}
\title{Format a pq_palette for printing}
\usage{
\method{format}{pq_palette}(x, ...)
}
\arguments{
\item{x}{(pq_palette, required) The palette to format.}

\item{...}{Ignored.}
}
\value{
A character vector of lines.
}
\description{
Format a pq_palette for printing
}
\author{
Adrien Taudière
}
\keyword{internal}
