% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_palette_class.R
\name{print.pq_palette}
\alias{print.pq_palette}
\title{Print a pq_palette}
\usage{
\method{print}{pq_palette}(x, ...)
}
\arguments{
\item{x}{(pq_palette, required) The palette to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a pq_palette
}
\author{
Adrien Taudière
}
\keyword{internal}
