% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_idest.R
\name{scale_color_idest_c}
\alias{scale_color_idest_c}
\title{IdEst continuous color scales for ggplot2}
\usage{
scale_color_idest_c(palette_name = "all_color_idest", direction = 1, ...)
}
\arguments{
\item{palette_name}{The name of the palette to use.
The available palette are c("all_color_idest", "ligth_color_idest",
"dark_color_idest", "Picabia", "Picasso", "Levine2", "Rattner", "Sidhu",
"Hokusai2", "Hokusai3"). See \link{idest_pal} for more details.}

\item{direction}{Direction of the palette. 1 for standard, -1 for reversed.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_color_gradientn]{ggplot2::scale_color_gradientn()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
IdEst continuous color scales for ggplot2
}
\author{
Adrien Taudière
}
