% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_idest.R
\name{scale_fill_idest_d}
\alias{scale_fill_idest_d}
\title{IdEst discrete fill scales for ggplot2}
\usage{
scale_fill_idest_d(
  palette_name = "all_color_idest",
  direction = 1,
  override_order = FALSE,
  ...
)
}
\arguments{
\item{palette_name}{The name of the palette to use.
The available palette are c("all_color_idest", "ligth_color_idest",
"dark_color_idest", "Picabia", "Picasso", "Levine2", "Rattner", "Sidhu",
"Hokusai2", "Hokusai3"). See \link{idest_pal} for more details.}

\item{direction}{Direction of the palette. 1 for standard, -1 for reversed.}

\item{override_order}{Logical (default FALSE),
whether to override the order of the palette.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
IdEst discrete fill scales for ggplot2
}
\author{
Adrien Taudière
}
