% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pq_discrete.R
\name{scale_fill_pq_discrete}
\alias{scale_fill_pq_discrete}
\title{ggplot2 discrete fill scale using the IdEst palette family}
\usage{
scale_fill_pq_discrete(
  palette = "Hokusai3",
  direction = 1,
  override_order = FALSE,
  ...
)
}
\arguments{
\item{palette}{(character, default "Hokusai3") The IdEst palette
to use. Must be one of the palettes in \link{idest_pal}.}

\item{direction}{(integer, default 1) Set to -1 to reverse the
palette.}

\item{override_order}{(logical, default FALSE) Override the curated
perceptual ordering of the palette.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} discrete fill scale.
}
\description{
ggplot2 discrete fill scale using the IdEst palette family
}
\examples{
\donttest{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_pq_discrete("Levine2")
}
}
\seealso{
\code{\link[=scale_color_pq_discrete]{scale_color_pq_discrete()}} for the matching colour scale.
}
\author{
Adrien Taudière
}
