% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotpq-package.R
\docType{package}
\name{ggplotpq-package}
\alias{ggplotpq}
\alias{ggplotpq-package}
\title{ggplotpq: 'ggplot2' Visualisation Helpers for 'phyloseq' Objects}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A 'ggplot2'-based visualisation layer for 'phyloseq' objects within the 'pqverse' ecosystem. Provides pure-ggplot2 helpers (geoms, themes, scales, color/palette management) and single-phyloseq ggplot wrappers migrated from 'MiscMetabar' and 'comparpq'. Pure rendering only: analysis methods that produce a plot as a side-effect belong in 'netaipq'; cross-package comparators belong in 'comparpq'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adrientaudiere.github.io/ggplotpq/}
  \item \url{https://github.com/adrientaudiere/ggplotpq}
  \item Report bugs at \url{https://github.com/adrientaudiere/ggplotpq/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]
}

}
\keyword{internal}
